requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
apscheduler>=3.10.0
pyyaml>=6.0.0
